Refactor of model and controller into separate file or article
