+++ controllers/topic-controller.js
@@ -1,9 +1,6 @@
 const {
   selectTopics,
-  selectArticles,
-  selectArticleById,
   selectUsers,
-  setArticleById,
 } = require("../models/topic-model.js");
 
 const getTopics = (request, response) => {
@@ -12,23 +9,6 @@ const getTopics = (request, response) => {
   });
 };
 
-const getArticles = (request, response) => {
-  selectArticles().then(({ rows: articles }) => {
-    response.status(200).send({ articles });
-  });
-};
-
-const getArticleById = (request, response, next) => {
-  const { article_id } = request.params;
-
-  selectArticleById(article_id)
-    .then((article) => {
-      response.status(200).send({ article });
-    })
-    .catch((err) => {
-      next(err);
-    });
-};
 
 const getUsers = (request, response) => {
   selectUsers().then((users) => {
@@ -36,25 +16,8 @@ const getUsers = (request, response) => {
   });
 };
 
-const updateArticleById = (request, response, next) => {
-
-  const { article_id } = request.params;
-  const { inc_votes } = request.body;
-
-  setArticleById(inc_votes, article_id)
-    .then((article) => {
-      response.status(200).send({ article });
-    })
-    .catch((err) => {
-      next(err);
-    });
-};
 
 module.exports = {
   getTopics,
   getUsers,
-  getUsers,
-  getArticles,
-  getArticleById,
-  updateArticleById,
 };
